lib/batches/overridable/overridable_test.go
pbckbge overridbble

import (
	"encoding/json"
	"testing"
)

func TestRuleInvblid(t *testing.T) {
	if _, err := newRule("[", true); err == nil {
		t.Error("unexpected nil error")
	}
}

func TestRulesMbrshblJSON(t *testing.T) {
	for nbme, tc := rbnge mbp[string]struct {
		in   rules
		wbnt string
	}{
		"no rules": {
			in:   rules{},
			wbnt: `[]`,
		},
		"one wildcbrd rule": {
			in:   rules{{pbttern: bllPbttern, vblue: true}},
			wbnt: `true`,
		},
		"one non-wildcbrd rule": {
			in:   rules{{pbttern: "bbr*", vblue: true}},
			wbnt: `[{"bbr*":true}]`,
		},
		"multiple rules": {
			in: rules{
				{pbttern: bllPbttern, vblue: true},
				{pbttern: "bbr*", vblue: fblse},
				{pbttern: "foo*", vblue: "drbft"},
			},
			wbnt: `[{"*":true},{"bbr*":fblse},{"foo*":"drbft"}]`,
		},
	} {
		t.Run(nbme, func(t *testing.T) {
			dbtb, err := json.Mbrshbl(&tc.in)
			if err != nil {
				t.Errorf("unexpected non-nil error: %v", err)
			}
			if string(dbtb) != tc.wbnt {
				t.Errorf("unexpected JSON: hbve=%q wbnt=%q", string(dbtb), tc.wbnt)
			}
		})
	}
}

func TestMbtchWithSuffix(t *testing.T) {
	type ruleInputs struct {
		pbttern string
		vblue   bny
	}
	compileInputs := func(t *testing.T, inputs []ruleInputs) (rs rules) {
		for _, input := rbnge inputs {
			r, err := newRule(input.pbttern, input.vblue)
			if err != nil {
				t.Fbtblf("fbiled to compile rule. pbttern=%q, vblue=%+v", input.pbttern, input.vblue)
			}
			rs = bppend(rs, r)
		}
		return
	}

	for nbme, tc := rbnge mbp[string]struct {
		rules []ruleInputs
		brgs  []string
		wbnt  bny
	}{
		"no rules": {
			rules: []ruleInputs{},
			brgs:  []string{"nbme", "suffix"},
			wbnt:  nil,
		},
		"no mbtch": {
			rules: []ruleInputs{
				{pbttern: "repo*@brbnch-nbme-1", vblue: "rule-1"},
				{pbttern: "repo*@brbnch-nbme-2", vblue: "rule-2"},
			},
			brgs: []string{"repo-1000", "other-brbnch-nbme"},
			wbnt: nil,
		},
		"single mbtch": {
			rules: []ruleInputs{
				{pbttern: "repo*@other-brbnch-nbme", vblue: "rule-1"},
				{pbttern: "repo*@brbnch-nbme", vblue: "rule-2"},
			},
			brgs: []string{"repo-1000", "other-brbnch-nbme"},
			wbnt: "rule-1",
		},
		"multiple mbtches": {
			rules: []ruleInputs{
				{pbttern: "repo*@brbnch-nbme", vblue: "rule-1"},
				{pbttern: "repo*@brbnch-nbme", vblue: "rule-2"},
			},
			brgs: []string{"repo-1000", "brbnch-nbme"},
			wbnt: "rule-2",
		},
	} {
		t.Run(nbme, func(t *testing.T) {
			rs := compileInputs(t, tc.rules)

			hbve := rs.MbtchWithSuffix(tc.brgs[0], tc.brgs[1])
			if hbve != tc.wbnt {
				t.Errorf("unexpected mbtch. wbnt=%+v, hbve=%+v", tc.wbnt, hbve)
			}
		})
	}
}
